Docs: Authenticate user using github and validation user in backend

+++ server/src/routes/memories.ts
@@ -3,8 +3,17 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
+    const { sub } = request.user
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -19,17 +28,23 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
-    const memory = prisma.memory.findUniqueOrThrow({
+    const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     })
+
+    const { sub } = request.user
+
+    if (!memory.isPublic && memory.userId !== sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -42,19 +57,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
 
     const { content, coverURL, isPublic } = bodySchema.parse(request.body)
+    const { sub } = request.user
     const memory = await prisma.memory.create({
       data: {
         content,
         coverURL,
         isPublic,
-        userId: '6d0207e5-84bf-4cf6-a523-0f9979701ab9',
+        userId: sub,
       },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -69,7 +85,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverURL, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findFirstOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    const { sub } = request.user
+
+    if (memory.userId !== sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -83,12 +111,24 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
-
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findFirstOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    const { sub } = request.user
+
+    if (memory.userId !== sub) {
+      return reply.status(401).send()
+    }
+
     await prisma.memory.delete({
       where: {
         id,
